Rename IKarmaLogger and friends to ITestLogger
