+++ KarmaTestAdapter/KarmaTestAdapter.csproj
@@ -96,15 +96,15 @@
     <Compile Include="Karma\KarmaServer.cs" />
     <Compile Include="Karma\KarmaSpec.cs" />
     <Compile Include="Karma\KarmaStopCommand.cs" />
-    <Compile Include="Logging\Extensions.cs" />
-    <Compile Include="Logging\IKarmaLogger.cs" />
-    <Compile Include="Logging\KarmaExtensibilityLogger.cs" />
-    <Compile Include="Logging\KarmaFileLogger.cs" />
-    <Compile Include="Logging\KarmaLogger.cs" />
-    <Compile Include="Logging\KarmaLoggerBase.cs" />
-    <Compile Include="Logging\KarmaLogLevel.cs" />
-    <Compile Include="Logging\KarmaServerLogger.cs" />
-    <Compile Include="Logging\KarmaTestMessageLogger.cs" />
+    <Compile Include="Logging\TestLoggerExtensions.cs" />
+    <Compile Include="Logging\ITestLogger.cs" />
+    <Compile Include="Logging\ExtensibilityTestLogger.cs" />
+    <Compile Include="Logging\FileTestLogger.cs" />
+    <Compile Include="Logging\TestLogger.cs" />
+    <Compile Include="Logging\TestLoggerBase.cs" />
+    <Compile Include="Logging\TestLogLevel.cs" />
+    <Compile Include="Logging\KarmaServerTestLogger.cs" />
+    <Compile Include="Logging\TestMessageTestLogger.cs" />
     <Compile Include="Helpers\PathUtils.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Helpers\TaskCompletionSourceExtensions.cs" />
