+++ KarmaTestAdapterTests/TestKarmaLogger.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 
 namespace KarmaTestAdapterTests
 {
-    public class TestKarmaLogger : KarmaLoggerBase
+    public class TestKarmaLogger : TestLoggerBase
     {
         public TestKarmaLogger(Action<string> logMessage)
         {
@@ -16,7 +16,7 @@ namespace KarmaTestAdapterTests
 
         public Action<string> LogMessage { get; private set; }
 
-        public override void Log(KarmaLogLevel level, IEnumerable<string> phase, string message)
+        public override void Log(TestLogLevel level, IEnumerable<string> phase, string message)
         {
             LogMessage(FormatMessage(level, phase, message));
         }
